chore: jwt whitelist에 batch 관련 제거, 배치용 컨트롤러 엔드포인트 변경, 공통 모듈 버전 업데이트
