+++ build.gradle
@@ -101,8 +101,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
     // [CouponPop Module]
-    implementation 'com.couponpop:couponpop-security:0.0.1-SNAPSHOT'
-    implementation 'com.couponpop:couponpop-core:0.0.5-SNAPSHOT'
+    implementation 'com.couponpop:couponpop-security:0.0.3-SNAPSHOT'
+    implementation 'com.couponpop:couponpop-core:0.0.7-SNAPSHOT'
 
     // [Spring Cloud Starter]
     implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
