+++ src/main/resources/application.yml
@@ -48,7 +48,6 @@ jwt:
       - /swagger-resources/**
       - /webjars/**
       - /docs/**
-      - /internal/batch/**
 
 logging:
   level:
